refactor: [#1284] extract udp-tracker-server package
