+++ .github/workflows/deployment.yaml
@@ -76,3 +76,4 @@ jobs:
           cargo publish -p torrust-tracker-primitives
           cargo publish -p torrust-tracker-test-helpers
           cargo publish -p torrust-tracker-torrent-repository
+          cargo publish -p torrust-udp-tracker-server
