+++ Cargo.toml
@@ -50,7 +50,6 @@ dashmap = "6"
 derive_more = { version = "2", features = ["as_ref", "constructor", "from"] }
 figment = "0"
 futures = "0"
-futures-util = "0"
 hyper = "1"
 parking_lot = "0"
 percent-encoding = "2"
@@ -60,7 +59,6 @@ r2d2_sqlite = { version = "0", features = ["bundled"] }
 rand = "0"
 regex = "1"
 reqwest = { version = "0", features = ["json"] }
-ringbuf = "0"
 serde = { version = "1", features = ["derive"] }
 serde_bencode = "0"
 serde_bytes = "0"
@@ -76,9 +74,9 @@ torrust-server-lib = { version = "3.0.0-develop", path = "packages/server-lib" }
 torrust-tracker-api-core = { version = "3.0.0-develop", path = "packages/tracker-api-core" }
 torrust-tracker-clock = { version = "3.0.0-develop", path = "packages/clock" }
 torrust-tracker-configuration = { version = "3.0.0-develop", path = "packages/configuration" }
-torrust-tracker-located-error = { version = "3.0.0-develop", path = "packages/located-error" }
 torrust-tracker-primitives = { version = "3.0.0-develop", path = "packages/primitives" }
 torrust-tracker-torrent-repository = { version = "3.0.0-develop", path = "packages/torrent-repository" }
+torrust-udp-tracker-server = { version = "3.0.0-develop", path = "packages/udp-tracker-server" }
 tower = { version = "0", features = ["timeout"] }
 tower-http = { version = "0", features = ["compression-full", "cors", "propagate-header", "request-id", "trace"] }
 tracing = "0"
