+++ packages/axum-http-tracker-server/src/v1/handlers/mod.rs
@@ -1,7 +1,4 @@
 //! Axum [`handlers`](axum#handlers) for the HTTP server.
-//!
-//! Refer to the generic [HTTP server documentation](crate::servers::http) for
-//! more information about the HTTP tracker.
 pub mod announce;
 pub mod common;
 pub mod health_check;
