+++ packages/axum-http-tracker-server/src/v1/mod.rs
@@ -1,7 +1,4 @@
 //! HTTP server implementation for the `v1` API.
-//!
-//! Refer to the generic [HTTP server documentation](crate::servers::http) for
-//! more information about the endpoints and their usage.
 pub mod extractors;
 pub mod handlers;
 pub mod routes;
