+++ packages/udp-tracker-server/README.md
@@ -1,0 +1,11 @@
+# Torrust UDP Tracker
+
+The Torrust Bittorrent UDP tracker.
+
+## Documentation
+
+[Crate documentation](https://docs.rs/torrust-udp-tracker-server).
+
+## License
+
+The project is licensed under the terms of the [GNU AFFERO GENERAL PUBLIC LICENSE](./LICENSE).
