+++ src/bootstrap/jobs/udp_tracker.rs
@@ -12,10 +12,9 @@ use bittorrent_udp_tracker_core::container::UdpTrackerCoreContainer;
 use bittorrent_udp_tracker_core::UDP_TRACKER_LOG_TARGET;
 use tokio::task::JoinHandle;
 use torrust_server_lib::registar::ServiceRegistrationForm;
+use torrust_udp_tracker_server::server::spawner::Spawner;
+use torrust_udp_tracker_server::server::Server;
 use tracing::instrument;
-
-use crate::servers::udp::server::spawner::Spawner;
-use crate::servers::udp::server::Server;
 
 /// It starts a new UDP server with the provided configuration.
 ///
