+++ src/servers/mod.rs
@@ -1,3 +1,2 @@
 //! Servers. Services that can be started and stopped.
 pub mod apis;
-pub mod udp;
