+++ tests/servers/health_check_api/contract.rs
@@ -14,7 +14,7 @@ async fn health_check_endpoint_should_return_status_ok_when_there_is_no_services
 
     let env = Started::new(&configuration.health_check_api.into(), Registar::default()).await;
 
-    let response = get(&format!("http://{}/health_check", env.state.binding)).await;
+    let response = get(&format!("http://{}/health_check", env.state.binding)).await; // DevSkim: ignore DS137138
 
     assert_eq!(response.status(), 200);
     assert_eq!(response.headers().get("content-type").unwrap(), "application/json");
@@ -54,7 +54,7 @@ mod api {
             let config = configuration.health_check_api.clone();
             let env = Started::new(&config.into(), registar).await;
 
-            let response = get(&format!("http://{}/health_check", env.state.binding)).await;
+            let response = get(&format!("http://{}/health_check", env.state.binding)).await; // DevSkim: ignore DS137138
 
             assert_eq!(response.status(), 200);
             assert_eq!(response.headers().get("content-type").unwrap(), "application/json");
@@ -76,7 +76,7 @@ mod api {
             assert_eq!(
                 details.info,
                 format!(
-                    "checking api health check at: http://{}/api/health_check",
+                    "checking api health check at: http://{}/api/health_check", // DevSkim: ignore DS137138
                     service.bind_address()
                 )
             );
@@ -105,7 +105,7 @@ mod api {
             let config = configuration.health_check_api.clone();
             let env = Started::new(&config.into(), registar).await;
 
-            let response = get(&format!("http://{}/health_check", env.state.binding)).await;
+            let response = get(&format!("http://{}/health_check", env.state.binding)).await; // DevSkim: ignore DS137138
 
             assert_eq!(response.status(), 200);
             assert_eq!(response.headers().get("content-type").unwrap(), "application/json");
@@ -131,7 +131,7 @@ mod api {
             );
             assert_eq!(
                 details.info,
-                format!("checking api health check at: http://{binding}/api/health_check")
+                format!("checking api health check at: http://{binding}/api/health_check") // DevSkim: ignore DS137138
             );
 
             env.stop().await.expect("it should stop the service");
@@ -164,7 +164,7 @@ mod http {
             let config = configuration.health_check_api.clone();
             let env = Started::new(&config.into(), registar).await;
 
-            let response = get(&format!("http://{}/health_check", env.state.binding)).await;
+            let response = get(&format!("http://{}/health_check", env.state.binding)).await; // DevSkim: ignore DS137138
 
             assert_eq!(response.status(), 200);
             assert_eq!(response.headers().get("content-type").unwrap(), "application/json");
@@ -185,7 +185,7 @@ mod http {
             assert_eq!(
                 details.info,
                 format!(
-                    "checking http tracker health check at: http://{}/health_check",
+                    "checking http tracker health check at: http://{}/health_check", // DevSkim: ignore DS137138
                     service.bind_address()
                 )
             );
@@ -214,7 +214,7 @@ mod http {
             let config = configuration.health_check_api.clone();
             let env = Started::new(&config.into(), registar).await;
 
-            let response = get(&format!("http://{}/health_check", env.state.binding)).await;
+            let response = get(&format!("http://{}/health_check", env.state.binding)).await; // DevSkim: ignore DS137138
 
             assert_eq!(response.status(), 200);
             assert_eq!(response.headers().get("content-type").unwrap(), "application/json");
@@ -240,7 +240,7 @@ mod http {
             );
             assert_eq!(
                 details.info,
-                format!("checking http tracker health check at: http://{binding}/health_check")
+                format!("checking http tracker health check at: http://{binding}/health_check") // DevSkim: ignore DS137138
             );
 
             env.stop().await.expect("it should stop the service");
@@ -273,7 +273,7 @@ mod udp {
             let config = configuration.health_check_api.clone();
             let env = Started::new(&config.into(), registar).await;
 
-            let response = get(&format!("http://{}/health_check", env.state.binding)).await;
+            let response = get(&format!("http://{}/health_check", env.state.binding)).await; // DevSkim: ignore DS137138
 
             assert_eq!(response.status(), 200);
             assert_eq!(response.headers().get("content-type").unwrap(), "application/json");
@@ -320,7 +320,7 @@ mod udp {
             let config = configuration.health_check_api.clone();
             let env = Started::new(&config.into(), registar).await;
 
-            let response = get(&format!("http://{}/health_check", env.state.binding)).await;
+            let response = get(&format!("http://{}/health_check", env.state.binding)).await; // DevSkim: ignore DS137138
 
             assert_eq!(response.status(), 200);
             assert_eq!(response.headers().get("content-type").unwrap(), "application/json");
